refactor: Board, Comment Mapper 재정의

+++ backend/src/main/java/com/ssafy/enjoytrip/board/model/dto/Board.java
@@ -18,7 +18,7 @@ public class Board {
     private String subject;
     private String content;
 
-    private Integer hit;
+    private Long hit;
     private String currentUpdate;
 
     private BoardType boardType;
